add feature carousel, and try it

+++ routes/web.php
@@ -19,7 +19,7 @@ use Illuminate\Http\Client\Request;
 // Route::view('/', 'welcome');
 
 Route::get('/', 'PostController@index');
-Route::get('posts/{post:slug}', 'PostController@show');
+Route::get('posts/{post:slug}', 'PostController@show')->name('showpost');
 
 Route::view('layanan', 'layanan');
 Route::view('tentangkami', 'tentangkami');
@@ -40,16 +40,26 @@ Route::get(
 
 Route::middleware('auth')->group(function () {
 
-    Route::get('admin', 'adminController@index')->name('admin');
-    Route::get('create', 'adminController@create')->name('create');
-    Route::post('store', 'adminController@store');
+    Route::get('admin', 'adminController@pengalihanAdmin');
+    //post
+    Route::get('admin/posts', 'adminController@index')->name('admin.posts');
+    Route::get('admin/posts/create', 'adminController@create')->name('create');
+    Route::post('admin/posts/store', 'adminController@store');
+    //update post
+    Route::get('admin/posts/{post:slug}/edit', 'adminController@edit');
+    Route::patch('admin/posts/{post:slug}/edit', 'adminController@update');
+    //delete post
+    Route::delete('admin/posts/{post:slug}/delete', 'adminController@destroy');
 
-    //update
-    Route::get('posts/{post:slug}/edit', 'adminController@edit');
-    Route::patch('posts/{post:slug}/edit', 'adminController@update');
-
-    //delete
-    Route::delete('posts/{post:slug}/delete', 'adminController@destroy');
+    //slider
+    Route::get('admin/sliders', 'sliderController@index');
+    Route::get('admin/sliders/create', 'sliderController@create');
+    Route::post('admin/sliders/store', 'sliderController@store');
+    //update slider
+    Route::get('admin/sliders/{slider:id}/edit', 'sliderController@edit');
+    Route::patch('admin/sliders/{slider:id}/update', 'sliderController@update');
+    //delete post
+    Route::delete('admin/sliders/{slider:id}/delete', 'sliderController@destroy');
 });
 
 Auth::routes([
